+++ boards/aerotenna/ocpoc/src/CMakeLists.txt
@@ -32,6 +32,7 @@
 ############################################################################
 
 add_library(drivers_board
+	board_pwm_out.cpp
 	i2c.cpp
 	spi.cpp
 )
