+++ boards/beaglebone/blue/src/CMakeLists.txt
@@ -32,6 +32,7 @@
 ############################################################################
 
 px4_add_library(drivers_board
+	board_pwm_out.cpp
 	i2c.cpp
 	init.cpp
 	spi.cpp
