+++ boards/beaglebone/blue/src/board_pwm_out.cpp
@@ -1,6 +1,6 @@
 /****************************************************************************
  *
- *   Copyright (c) 2017 PX4 Development Team. All rights reserved.
+ *   Copyright (c) 2018 PX4 Development Team. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -31,27 +31,54 @@
  *
  ****************************************************************************/
 
-#pragma once
+#ifndef MODULE_NAME
+#define MODULE_NAME "bbblue_pwm_out"
+#endif
 
-#include <stdint.h>
+#include <fcntl.h>
+#include <errno.h>
+#include <px4_platform_common/log.h>
 
-namespace linux_pwm_out
+#include <robotcontrol.h>
+#include <board_config.h>
+
+#include "board_pwm_out.h"
+
+using namespace pwm_out;
+
+BBBlueRcPWMOut::BBBlueRcPWMOut(int max_num_outputs) : _num_outputs(max_num_outputs)
 {
+	if (_num_outputs > MAX_NUM_PWM) {
+		PX4_WARN("number of outputs too large. Setting to %i", MAX_NUM_PWM);
+		_num_outputs = MAX_NUM_PWM;
+	}
+}
 
-/**
- ** class PWMOutBase
- * common abstract PWM output base class
- */
-class PWMOutBase
+BBBlueRcPWMOut::~BBBlueRcPWMOut()
 {
-public:
+	rc_cleaning();
+}
 
-	virtual ~PWMOutBase() {}
+int BBBlueRcPWMOut::init()
+{
+	rc_init();
 
-	virtual int init() = 0;
+	return 0;
+}
 
-	virtual int send_output_pwm(const uint16_t *pwm, int num_outputs) = 0;
-};
+int BBBlueRcPWMOut::send_output_pwm(const uint16_t *pwm, int num_outputs)
+{
+	if (num_outputs > _num_outputs) {
+		num_outputs = _num_outputs;
+	}
 
+	int ret = 0;
 
-} /* namespace rpi_pwm_out */
+	// pwm[ch] is duty_cycle in us
+	for (int ch = 0; ch < num_outputs; ++ch) {
+		ret += rc_servo_send_pulse_us(ch + 1, pwm[ch]); // converts to 1-based channel #
+	}
+
+	return ret;
+}
+
