+++ boards/px4/raspberrypi/default.cmake
@@ -23,7 +23,6 @@ px4_add_board(
 		gps
 		#imu # all available imu drivers
 		imu/invensense/mpu9250
-		linux_pwm_out
 		#magnetometer # all available magnetometer drivers
 		magnetometer/hmc5883
 		pca9685_pwm_out
