+++ posix-configs/bbblue/px4.config
@@ -83,8 +83,8 @@ sleep 1
 #rc_input start -d /dev/ttyS4
 
 # default: etc/mixers/quad_x.main.mix, 8 output channels
-linux_pwm_out start -p bbblue_rc -m etc/mixers/quad_x.main.mix
-#linux_pwm_out start -p bbblue_rc -m etc/mixers/AETRFG.main.mix
+linux_pwm_out start -m etc/mixers/quad_x.main.mix
+#linux_pwm_out start -m etc/mixers/AETRFG.main.mix
 
 logger start -t -b 200
