+++ posix-configs/ocpoc/px4.config
@@ -45,6 +45,6 @@ mavlink stream -d /dev/ttyPS1 -s ATTITUDE -r 50
 
 rc_input start -d /dev/ttyS2
 
-linux_pwm_out start -p ocpoc_mmap -m etc/mixers/quad_x.main.mix
+linux_pwm_out start -m etc/mixers/quad_x.main.mix
 logger start -t -b 200
 mavlink boot_complete
