+++ src/drivers/linux_pwm_out/CMakeLists.txt
@@ -36,11 +36,7 @@ px4_add_module(
 	MAIN linux_pwm_out
 	COMPILE_FLAGS
 	SRCS
-		PCA9685.cpp
-		navio_sysfs.cpp
 		linux_pwm_out.cpp
-		ocpoc_mmap.cpp
-		bbblue_pwm_rc.cpp
 	DEPENDS
 		output_limit
 	)
